[identity] create service template
